Android Sensor based game
Developed a barrel racing game in a group of three using various android
sensors wherein the user is challenged to complete the game in the
minimum possible time. The user has to overcome all the obstacles in
this attempt to complete the entire race track. The best scores are
updated with each race.
Technology used: Android ADT